# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ghar Ka Mali API",
    "version": "2.0.0",
    "description": "Complete gardening platform API — 83 endpoints"
  },
  "servers": [
    { "url": "http://YOUR_LIGHTSAIL_IP:3000/api", "description": "Production (Lightsail)" },
    { "url": "http://localhost:3000/api", "description": "Local Development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
...
  "paths": {
    "/auth/send-otp":           { "post": { "tags":["Auth"], "summary":"Send OTP to phone", "requestBody":{"content":{"application/json":{"schema":{"properties":{"phone":{"type":"string","example":"9876543210"}}}}}}, "responses":{"200":{"description":"OTP sent (static: 123456 in dev)"}} } },
    "/auth/verify-otp":         { "post": { "tags":["Auth"], "summary":"Verify OTP → login or register customer", "requestBody":{"content":{"application/json":{"schema":{"properties":{"phone":{"type":"string"},"otp":{"type":"string","example":"123456"},"name":{"type":"string"}}}}}}, "responses":{"200":{"description":"JWT token + user"}} } },
    "/auth/admin-login":        { "post": { "tags":["Auth"], "summary":"Admin/Supervisor password login", "requestBody":{"content":{"application/json":{"schema":{"properties":{"phone":{"type":"string"},"password":{"type":"string"}}}}}}, "responses":{"200":{"description":"JWT token"}} } },
    "/auth/gardener-login":     { "post": { "tags":["Auth"], "summary":"Gardener OTP login", "requestBody":{"content":{"application/json":{"schema":{"properties":{"phone":{"type":"string"},"otp":{"type":"string"}}}}}}, "responses":{"200":{"description":"JWT + gardener profile"}} } },
    "/auth/gardener-register":  { "post": { "tags":["Auth"], "summary":"Gardener registration (multipart)", "requestBody":{"content":{"multipart/form-data":{"schema":{"properties":{"name":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"experience_years":{"type":"integer"},"bio":{"type":"string"},"service_zone_ids":{"type":"array","items":{"type":"integer"}},"profile_image":{"type":"string","format":"binary"},"id_proof":{"type":"string","format":"binary"}}}}}}, "responses":{"201":{"description":"Submitted, awaiting approval"}} } },
    "/auth/profile":            { "get":{"tags":["Auth"],"summary":"Get current user profile","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"User + gardener profile if applicable"}}}, "put":{"tags":["Auth"],"summary":"Update profile","security":[{"bearerAuth":[]}],"requestBody":{"content":{"multipart/form-data":{"schema":{"properties":{"name":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"profile_image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Updated user"}}} },
...
    "/payments/check-serviceability": { "get": { "tags":["Payments"], "summary":"Check if lat/lng is in serviceable zone", "parameters":[{"name":"latitude","in":"query","required":true,"schema":{"type":"number"}},{"name":"longitude","in":"query","required":true,"schema":{"type":"number"}}], "responses":{"200":{"description":"serviceable: bool, zones: []"}} } },
    "/payments/initiate":       { "post": { "tags":["Payments"], "summary":"Initiate PayU payment", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"type":{"type":"string","enum":["booking","subscription","wallet_topup"]},"booking_id":{"type":"integer"},"subscription_id":{"type":"integer"},"amount":{"type":"number"}}}}}}, "responses":{"200":{"description":"PayU params + payment_id"}} } },
...
    "/payments/status/{txnid}": { "get": { "tags":["Payments"], "summary":"Check payment status by transaction ID", "security":[{"bearerAuth":[]}], "parameters":[{"name":"txnid","in":"path","required":true,"schema":{"type":"string"}}], "responses":{"200":{"description":"Payment record"}} } },
    "/payments/my":             { "get": { "tags":["Payments"], "summary":"Get my payment history", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"Paginated payments"}} } },
    "/payments/wallet-topup":   { "post": { "tags":["Payments"], "summary":"Initiate wallet top-up via PayU", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"amount":{"type":"number","minimum":100}}}}}}, "responses":{"200":{"description":"PayU params"}} } },
    "/payments/reschedule":     { "post": { "tags":["Bookings"], "summary":"Reschedule a booking", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"booking_id":{"type":"integer"},"new_date":{"type":"string","format":"date"},"new_time":{"type":"string"}}}}}}, "responses":{"200":{"description":"Rescheduled booking"}} } },
    "/bookings":                { "post": { "tags":["Bookings"], "summary":"Create on-demand booking", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"required":["zone_id","scheduled_date","service_address","service_latitude","service_longitude"],"properties":{"zone_id":{"type":"integer"},"scheduled_date":{"type":"string","format":"date"},"scheduled_time":{"type":"string"},"service_address":{"type":"string"},"service_latitude":{"type":"number"},"service_longitude":{"type":"number"},"plant_count":{"type":"integer"},"preferred_gardener_id":{"type":"integer"},"customer_notes":{"type":"string"}}}}}}, "responses":{"201":{"description":"Booking with OTP"}} } },
    "/bookings/my":             { "get": { "tags":["Bookings"], "summary":"Get my bookings (customer)", "security":[{"bearerAuth":[]}], "parameters":[{"name":"status","in":"query","schema":{"type":"string"}},{"name":"page","in":"query","schema":{"type":"integer"}}], "responses":{"200":{"description":"Paginated bookings"}} } },
    "/bookings/{id}":           { "get": { "tags":["Bookings"], "summary":"Get booking detail", "security":[{"bearerAuth":[]}], "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}], "responses":{"200":{"description":"Full booking with gardener & zone"}} } },
    "/bookings/verify-otp":     { "post": { "tags":["Gardener"], "summary":"Verify visit OTP to start job", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"booking_id":{"type":"integer"},"otp":{"type":"string"}}}}}}, "responses":{"200":{"description":"Visit started"}} } },
    "/bookings/status":         { "put": { "tags":["Gardener"], "summary":"Update booking status + work proof images", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"multipart/form-data":{"schema":{"properties":{"booking_id":{"type":"integer"},"status":{"type":"string","enum":["en_route","arrived","in_progress","completed","failed"]},"gardener_notes":{"type":"string"},"extra_plants":{"type":"integer"},"before_image":{"type":"string","format":"binary"},"after_image":{"type":"string","format":"binary"}}}}}}, "responses":{"200":{"description":"Status updated"}} } },
    "/bookings/rate":           { "post": { "tags":["Bookings"], "summary":"Rate a completed booking (1-5 stars)", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"booking_id":{"type":"integer"},"rating":{"type":"integer","minimum":1,"maximum":5},"review":{"type":"string"}}}}}}, "responses":{"200":{"description":"Rated, gardener avg updated"}} } },
    "/bookings/cancel":         { "post": { "tags":["Bookings"], "summary":"Cancel a booking", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"booking_id":{"type":"integer"},"reason":{"type":"string"}}}}}}, "responses":{"200":{"description":"Cancelled"}} } },
    "/bookings/location":       { "post": { "tags":["Gardener"], "summary":"Update gardener GPS location", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"latitude":{"type":"number"},"longitude":{"type":"number"},"booking_id":{"type":"integer"}}}}}}, "responses":{"200":{"description":"Location saved + Socket.io broadcast"}} } },
    "/bookings/track/{booking_id}": { "get": { "tags":["Bookings"], "summary":"Get gardener current location for tracking", "security":[{"bearerAuth":[]}], "parameters":[{"name":"booking_id","in":"path","required":true,"schema":{"type":"integer"}}], "responses":{"200":{"description":"lat/lng + last_updated"}} } },
    "/bookings/gardener/jobs":  { "get": { "tags":["Gardener"], "summary":"Get gardener's assigned jobs", "security":[{"bearerAuth":[]}], "parameters":[{"name":"status","in":"query","schema":{"type":"string"}},{"name":"date","in":"query","schema":{"type":"string","format":"date"}},{"name":"page","in":"query","schema":{"type":"integer"}}], "responses":{"200":{"description":"Jobs list"}} } },
    "/bookings/gardener/earnings": { "get": { "tags":["Gardener"], "summary":"Gardener earnings breakdown by period", "security":[{"bearerAuth":[]}], "parameters":[{"name":"period","in":"query","schema":{"type":"string","enum":["daily","weekly","monthly"]}}], "responses":{"200":{"description":"Breakdown + totals"}} } },
    "/bookings/{id}/addons":    { "get":{"tags":["Add-Ons"],"summary":"Get add-ons for a booking","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Add-ons list"}}}, "post":{"tags":["Add-Ons"],"summary":"Add extra services to a booking","security":[{"bearerAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"addon_ids":{"type":"array","items":{"type":"object","properties":{"addon_id":{"type":"integer"},"quantity":{"type":"integer"}}}}}}}}},"responses":{"200":{"description":"Add-ons attached, total updated"}}} },
    "/subscriptions":           { "post": { "tags":["Subscriptions"], "summary":"Purchase a subscription plan", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"required":["plan_id","service_address","service_latitude","service_longitude"],"properties":{"plan_id":{"type":"integer"},"zone_id":{"type":"integer"},"service_address":{"type":"string"},"service_latitude":{"type":"number"},"service_longitude":{"type":"number"},"plant_count":{"type":"integer"},"preferred_gardener_id":{"type":"integer"},"auto_renew":{"type":"boolean"}}}}}}, "responses":{"201":{"description":"Subscription + auto-scheduled visits"}} } },
    "/subscriptions/my":        { "get": { "tags":["Subscriptions"], "summary":"Get my subscriptions", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"Subscriptions with plan details"}} } },
    "/subscriptions/{id}/cancel": { "put": { "tags":["Subscriptions"], "summary":"Cancel subscription", "security":[{"bearerAuth":[]}], "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}], "responses":{"200":{"description":"Cancelled"}} } },
    "/plants/identify":         { "post": { "tags":["Plantopedia"], "summary":"AI plant identification from image", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"multipart/form-data":{"schema":{"properties":{"image":{"type":"string","format":"binary"}}}}}}, "responses":{"200":{"description":"Plant name, care guide, watering schedule, fertilizer tips"}} } },
    "/plants/history":          { "get": { "tags":["Plantopedia"], "summary":"Get my plant identification history", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"History list"}} } },
    "/blogs":                   { "get": { "tags":["Content"], "summary":"Get published blogs (SSG)", "parameters":[{"name":"page","in":"query","schema":{"type":"integer"}},{"name":"category","in":"query","schema":{"type":"string"}},{"name":"city_slug","in":"query","schema":{"type":"string"}}], "responses":{"200":{"description":"Paginated blogs"}} } },
    "/blogs/{slug}":            { "get": { "tags":["Content"], "summary":"Get blog post by slug", "parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}], "responses":{"200":{"description":"Full blog post, views incremented"}} } },
...
    "/cities/{slug}":           { "get": { "tags":["Content"], "summary":"Get city page by slug", "parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}], "responses":{"200":{"description":"City SEO page"}} } },
    "/notifications":           { "get": { "tags":["Notifications"], "summary":"Get my notifications", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"30 most recent"}} } },
    "/notifications/{id}/read": { "put": { "tags":["Notifications"], "summary":"Mark notification as read", "security":[{"bearerAuth":[]}], "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}], "responses":{"200":{"description":"Marked"}} } },
    "/complaints":              { "post":{"tags":["Complaints"],"summary":"Raise a complaint (customer)","security":[{"bearerAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"required":["type","description"],"properties":{"booking_id":{"type":"integer"},"type":{"type":"string","enum":["service_quality","late_arrival","no_show","rude_behavior","billing","damage","other"]},"description":{"type":"string"},"priority":{"type":"string","enum":["low","medium","high"]}}}}}},"responses":{"201":{"description":"Complaint raised, supervisor notified via WhatsApp"}}}, "get":{"tags":["Complaints"],"summary":"Get all complaints (admin/supervisor)","security":[{"bearerAuth":[]}],"parameters":[{"name":"status","in":"query","schema":{"type":"string"}},{"name":"priority","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Complaints list"}}} },
    "/complaints/my":           { "get": { "tags":["Complaints"], "summary":"Get my complaints (customer)", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"Customer complaints"}} } },
    "/complaints/stats":        { "get": { "tags":["Complaints"], "summary":"Complaint stats (admin/supervisor)", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"open/inReview/resolved/highPriority counts"}} } },
    "/complaints/{id}":         { "put": { "tags":["Complaints"], "summary":"Update complaint — assign/resolve (admin/supervisor)", "security":[{"bearerAuth":[]}], "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"status":{"type":"string"},"resolution_notes":{"type":"string"},"assigned_to":{"type":"integer"}}}}}}, "responses":{"200":{"description":"Updated, customer notified"}} } },
    "/supervisor/dashboard":    { "get": { "tags":["Supervisor"], "summary":"Supervisor team dashboard", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"Team gardeners + today stats"}} } },
    "/admin/dashboard":         { "get": { "tags":["Admin"], "summary":"Admin overview stats", "security":[{"bearerAuth":[]}], "responses":{"200":{"description":"Stats + recent bookings"}} } },
    "/admin/analytics":         { "get": { "tags":["Admin"], "summary":"Full analytics data", "security":[{"bearerAuth":[]}], "parameters":[{"name":"period","in":"query","schema":{"type":"string","default":"30"},"description":"Days to look back"},{"name":"zone_id","in":"query","schema":{"type":"integer"}}], "responses":{"200":{"description":"Revenue, zones, cities, users, plans, ratings, repeat customers"}} } },
    "/admin/utilization":       { "get": { "tags":["Admin"], "summary":"Gardener utilization rate report", "security":[{"bearerAuth":[]}], "parameters":[{"name":"period","in":"query","schema":{"type":"string","default":"30"}}], "responses":{"200":{"description":"Per-gardener utilization % + summary"}} } },
    "/admin/gardeners":         { "get":{"tags":["Admin"],"summary":"List gardeners with filter","security":[{"bearerAuth":[]}],"parameters":[{"name":"status","in":"query","schema":{"type":"string","enum":["pending","active","inactive",""]}},{"name":"search","in":"query","schema":{"type":"string"}},{"name":"city","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Paginated gardeners with profiles"}}} },
    "/admin/gardeners/approve": { "post": { "tags":["Admin"], "summary":"Approve gardener + assign supervisor", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"user_id":{"type":"integer"},"supervisor_id":{"type":"integer"}}}}}}, "responses":{"200":{"description":"Approved, WhatsApp sent"}} } },
    "/admin/gardeners/reject":  { "post": { "tags":["Admin"], "summary":"Reject gardener application", "security":[{"bearerAuth":[]}], "requestBody":{"content":{"application/json":{"schema":{"properties":{"user_id":{"type":"integer"},"reason":{"type":"string"}}}}}}, "responses":{"200":{"description":"Rejected"}} } },
...
    "/admin/zones/{id}":        { "put": { "tags":["Admin"], "summary":"Update zone", "security":[{"bearerAuth":[]}], "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}], "requestBody":{"content":{"application/json":{"schema":{}}}}, "responses":{"200":{"description":"Updated"}} } },
...
  }
}